requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
sentence-transformers==2.7.0
huggingface_hub==0.20.3
transformers==4.37.2
tokenizers==0.15.2
torch==2.2.0
nltk==3.8.1
pandas==2.3.1
pydantic>=2.0.0,<3.0.0